Added change parameters of bloc

+++ Assets/Scripts/Bloc.cs
@@ -6,7 +6,7 @@ public class Bloc
 {
     public int numBloc { get; }
     public int pot { get; set; }
-    public float temps { get; set; }
+    public int temps { get; set; }
 
     public Bloc(int numBloc)
     {
